% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IalsaSynthesis-package.r
\docType{package}
\name{IalsaSynthesis-package}
\alias{IalsaSynthesis-package}
\title{Synthesizing Information Across Collaborating Research}
\description{
Synthesizing information across collaborating research. Created for \href{https://www.ialsa.org/#team}{Integrative Analysis of Longitudinal Studies of Aging} (IALSA).
}
\note{
The release version will be available through \href{https://cran.r-project.org/package=IalsaSynthesis}{CRAN}.
The most recent development version is available through \href{https://github.com/IALSA/IalsaSynthesis}{GitHub}.
Please see the installation examples below.

If you're having trouble with the package, please install the development version.  If this doesn't solve
your problem, please create a \href{https://github.com/IALSA/IalsaSynthesis/issues}{new issue}, or email Will or Andriy.
}
\examples{
\dontrun{
# Install/update IalsaSynthesis with the release version from CRAN.
install.packages('IalsaSynthesis')

# Install/update IalsaSynthesis with the development version from GitHub
#install.packages('devtools') #Uncomment if `devtools` isn't installed already.
devtools::install_github('IALSA/IalsaSynthesis')
}
}
\references{
IALSA is funded through an NIH/NIA Program Project Grant (
\href{https://projectreporter.nih.gov/project_info_description.cfm?aid=8414933&icde=18870651&ddparam=&ddvalue=&ddsub=&cr=3&csb=default&cs=ASC}{P01AG043362}); 2013-2018) to Oregon Health & Science University
(Program Directors:
\href{https://www.ialsa.org/#team}{Scott Hofer},
\href{https://www.ialsa.org/#team}{Andrea Piccinin},
\href{https://www.ialsa.org/#team}{Jeffrey Kaye}, and
\href{https://www.ialsa.org/#team}{Diana Kuh})
and previously funded by NIH/NIA (\href{https://projectreporter.nih.gov/project_info_description.cfm?aid=7210005&icde=19139556&ddparam=&ddvalue=&ddsub=&cr=5&csb=default&cs=ASC}{R01AG026453}; 2007-2013) and CIHR (103284; 2010-2013).
}
\author{
\href{http://scholar.google.com/citations?user=ffsJTC0AAAAJ}{William Howard Beasley}
 \href{https://medicine.ouhsc.edu/Academic-Departments/Pediatrics}{Dept of Pediatrics}, Biomedical and Behavioral Methodology Core (\href{https://www.ouhsc.edu/bbmc/}{BBMC})


 \href{https://www.researchgate.net/profile/Andriy-Koval-2}{Andriy Koval}
 --Post Doc, \href{http://www.uvic.ca/}{University of Victoria}, \href{http://www.uvic.ca/socialsciences/psychology/}{Department of Psychology}
}
