% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\alias{extract_aic}
\alias{extract_bic}
\alias{extract_bic_adjusted}
\alias{extract_free_parameter_count}
\alias{extract_loglikelihood}
\alias{extract_output_filename}
\alias{extract_scaling_correction}
\title{Extract the values within model output files.}
\usage{
extract_output_filename(mplus_output)
}
\arguments{
\item{mplus_output}{Text containing model output.}
}
\value{
A \code{numeric} value corresponding to the desired quantity.
}
\description{
Functions that extract the values within model output files.
}
\examples{
library(IalsaSynthesis) #Load the package into the current R session.
}
\author{
Will Beasley
}

