% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\alias{extract_aic}
\alias{extract_bic}
\alias{extract_bic_adjusted}
\alias{extract_free_parameter_count}
\alias{extract_loglikelihood}
\alias{extract_output_filename}
\alias{extract_scalar_float}
\alias{extract_scalar_string}
\alias{extract_scaling_correction}
\title{Extract the values within model output files.}
\usage{
extract_scalar_string(regex, source)

extract_scalar_float(regex, source)

extract_output_filename(mplus_output,
  regex = "\\\\s+DATA:\\\\s+File = (.+);.*")

extract_free_parameter_count(mplus_output,
  regex = "Number of Free Parameters\\\\s+(\\\\d{1,})\\\\s+")

extract_loglikelihood(mplus_output,
  regex = "Loglikelihood\\\\s+H0 Value\\\\s+([-\\\\d\\\\.]+)\\\\s+")

extract_scaling_correction(mplus_output,
  regex = "\\\\s+H0 Scaling Correction Factor\\\\s+([-\\\\d\\\\.]+)\\\\s+for MLR\\\\s+")

extract_aic(mplus_output,
  regex = "Akaike \\\\(AIC\\\\)\\\\s+([-\\\\d\\\\.]+)\\\\s+")

extract_bic(mplus_output,
  regex = "Bayesian \\\\(BIC\\\\)\\\\s+([-\\\\d\\\\.]+)\\\\s+")

extract_bic_adjusted(mplus_output,
  regex = "\\\\s+Sample-Size Adjusted BIC\\\\s+([-\\\\d\\\\.]+)\\\\s+")
}
\arguments{
\item{regex}{Regular Expression pattern to capture and extract contents.}

\item{source}{Text to run the regex against.}

\item{mplus_output}{Text containing model output.  This should be the text read from the file (not a file path).}
}
\value{
A \code{numeric} value corresponding to the desired quantity.
}
\description{
Functions that extract the values within model output files.
}
\section{Functions}{
\itemize{
\item \code{extract_scalar_float}: Generalizable function to return a single numeric value.

\item \code{extract_output_filename}: Determine the path of the Mplus output file.

\item \code{extract_free_parameter_count}: Determine the number of free parameters for the model estimation.

\item \code{extract_loglikelihood}: Determine the log likelihood for the model estimation.

\item \code{extract_scaling_correction}: Determine the H0 Scaling Correction Factor for the model estimation.

\item \code{extract_aic}: Determine the AIC for the model estimation.

\item \code{extract_bic}: Determine the BIC for the model estimation.

\item \code{extract_bic_adjusted}: Determine the Sample-Size Adjusted BIC for the model estimation.
}}
\examples{
library(IalsaSynthesis) #Load the package into the current R session.
}
\author{
Will Beasley
}

