% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\alias{extract_scalar_string}
\alias{extract_scalar_float}
\alias{extract_named_wald}
\alias{extract_output_filename}
\alias{extract_free_parameter_count}
\alias{extract_loglikelihood}
\alias{extract_scaling_correction}
\alias{extract_aic}
\alias{extract_bic}
\alias{extract_bic_adjusted}
\title{Extract the values within model output files.}
\usage{
extract_scalar_string(regex, source)

extract_scalar_float(regex, source)

extract_named_wald(parameter_name, mplus_output)

extract_output_filename(mplus_output,
  regex = "\\\\bDATA:\\\\s+File = (.+);.*")

extract_free_parameter_count(mplus_output,
  regex = "\\\\bNumber of Free Parameters\\\\s+(\\\\d{1,})\\\\b")

extract_loglikelihood(mplus_output,
  regex = "\\\\bLoglikelihood\\\\s+H0 Value\\\\s+([-\\\\d\\\\.]+)\\\\b")

extract_scaling_correction(mplus_output,
  regex = "\\\\bH0 Scaling Correction Factor\\\\s+([-\\\\d\\\\.]+)\\\\s+for MLR\\\\b")

extract_aic(mplus_output,
  regex = "\\\\bAkaike \\\\(AIC\\\\)\\\\s+([-\\\\d\\\\.]+)\\\\b")

extract_bic(mplus_output,
  regex = "\\\\bBayesian \\\\(BIC\\\\)\\\\s+([-\\\\d\\\\.]+)\\\\b")

extract_bic_adjusted(mplus_output,
  regex = "\\\\bSample-Size Adjusted BIC\\\\s+([-\\\\d\\\\.]+)\\\\b")
}
\arguments{
\item{regex}{Regular Expression pattern to capture and extract contents.}

\item{source}{Text to run the regex against.}

\item{parameter_name}{Variable name in Mplus output to extract.}

\item{mplus_output}{Text containing model output.  This should be the text read from the file (not a file path).}
}
\value{
A \code{numeric} value corresponding to the desired quantity.
}
\description{
Functions that extract the values within model output files.
}
\section{Functions}{
\itemize{
\item \code{extract_scalar_float}: Generalizable function to return a single numeric value.

\item \code{extract_named_wald}: Determine the estimate, standard error, z-score, and two-tailed p-value of an estimate (ie, the results of a Wald test).

\item \code{extract_output_filename}: Determine the path of the Mplus output file.

\item \code{extract_free_parameter_count}: Determine the number of free parameters for the model estimation.

\item \code{extract_loglikelihood}: Determine the log likelihood for the model estimation.

\item \code{extract_scaling_correction}: Determine the H0 Scaling Correction Factor for the model estimation.

\item \code{extract_aic}: Determine the AIC for the model estimation.

\item \code{extract_bic}: Determine the BIC for the model estimation.

\item \code{extract_bic_adjusted}: Determine the Sample-Size Adjusted BIC for the model estimation.
}}

\examples{
library(IalsaSynthesis) #Load the package into the current R session.
}
\author{
Will Beasley
}
